requirements.txt
# requirements.txt (mlops-project root level)

# ========================================
# CORE DATA & ML
# ========================================
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.3.2
pyarrow==12.0.1
fairlearn==0.9.0

# ========================================
# ORCHESTRATION
# ========================================
apache-airflow==2.9.2
apache-airflow-providers-google==10.13.0
psycopg2-binary==2.9.9

# ========================================
# GOOGLE CLOUD PLATFORM
# ========================================
google-cloud-storage==2.12.0
google-cloud-bigquery==3.14.0
google-cloud-aiplatform>=1.70.0
google-re2==1.1.20240702


# ========================================
# DATA VERSIONING
# ========================================
dvc[gs]==3.51.0

# ========================================
# MODEL TRAINING SPECIFIC
# ========================================
duckdb==0.9.2

# ========================================
# UTILITIES
# ========================================
python-dotenv==1.0.0
chardet==5.2.0
requests==2.31.0

# ========================================
# TESTING
# ========================================
pytest==7.4.0

streamlit
plotly
# ========================================
# CI/CD
# ========================================
PyYAML==6.0.1

# ========================================
# To enable embedding - prompt drift detction
# ========================================
 google-generativeai
 